+++ packages/browser/src/Session.ts
@@ -217,14 +217,14 @@ export class Session extends EventEmitter {
       url
     );
     if (sessionInfo) {
+      this.info.isLoggedIn = sessionInfo.isLoggedIn;
+      this.info.webId = sessionInfo.webId;
+      this.info.sessionId = sessionInfo.sessionId;
       if (sessionInfo.isLoggedIn) {
         // The login event can only be triggered **after** the user has been
         // redirected from the IdP with access and ID tokens.
         this.emit("login");
       }
-      this.info.isLoggedIn = sessionInfo.isLoggedIn;
-      this.info.webId = sessionInfo.webId;
-      this.info.sessionId = sessionInfo.sessionId;
     }
     this.tokenRequestInProgress = false;
     return sessionInfo;