+++ packages/node/src/Session.ts
@@ -207,21 +207,19 @@ export class Session extends EventEmitter {
         );
 
         if (sessionInfo) {
+          this.info.isLoggedIn = sessionInfo.isLoggedIn;
+          this.info.webId = sessionInfo.webId;
+          this.info.sessionId = sessionInfo.sessionId;
           if (sessionInfo.isLoggedIn) {
             // The login event can only be triggered **after** the user has been
             // redirected from the IdP with access and ID tokens.
             this.emit("login");
           }
-
-          this.info.isLoggedIn = sessionInfo.isLoggedIn;
-          this.info.webId = sessionInfo.webId;
-          this.info.sessionId = sessionInfo.sessionId;
         }
       } finally {
         this.tokenRequestInProgress = false;
       }
     }
-
     return sessionInfo;
   };